Citiesearch/requirements/local.txt
aioredis==1.3.1
alabaster==0.7.12
appdirs==1.4.4
argon2-cffi==20.1.0
arrow==0.17.0
asgi-redis==1.4.3
asgiref==3.3.1
astroid==2.4.2
async-timeout==3.0.1
attrs==20.3.0
autobahn==21.2.1
Automat==20.2.0
Babel==2.9.0
backcall==0.2.0
beautifulsoup4==4.9.3
binaryornot==0.4.4
black==20.8b1
boto3==1.16.63
botocore==1.19.63
bs4==0.0.1
certifi==2020.11.8
cffi==1.14.4
cfgv==3.2.0
channels==3.0.3
channels-redis==3.2.0
chardet==3.0.4
click==7.1.2
Collectfast==2.2.0
constantly==15.1.0
cookiecutter==1.7.2
coverage==5.4
cryptography==3.3.1
cssselect==1.1.0
daphne==3.0.1
dataclasses==0.8
decorator==4.4.2
defusedxml==0.6.0
distlib==0.3.1
Django==3.0.11
django-allauth==0.44.0
django-anymail==8.1
django-appconf==1.0.4
django-compressor==2.4
django-coverage-plugin==1.8.0
django-crispy-forms==1.10.0
django-debug-toolbar==3.2
django-deferred-polymorph==0.4.0
django-environ==0.4.5
django-extensions==3.1.1
django-extra-views==0.13.0
django-geojson==3.1.0
django-leaflet==0.27.1
django-libsass==0.8
django-mathfilters==1.0.0
django-model-utils==4.1.1
django-price==0.2.1
django-prices==2.2.0
django-redis==4.12.1
django-storages==1.11.1
django-stubs==1.7.0
django-tagulous==1.1.0
djangorestframework==3.12.2
django_resized
docutils==0.16
enmerkar==0.7.1
factory-boy==3.2.0
fake-useragent==0.1.11
Faker==5.7.0
filelock==3.0.12
flake8==3.8.4
flake8-isort==4.0.0
gunicorn==20.0.4
hiredis==1.1.0
hyperlink==21.0.0
identify==1.5.13
idna==2.10
imagesize==1.2.0
importlib-metadata==3.4.0
importlib-resources==5.1.0
incremental==21.3.0
iniconfig==1.1.1
ipdb==0.13.4
ipython==7.16.1
ipython-genutils==0.2.0
isort==5.6.1
jedi==0.18.0
Jinja2==2.11.2
jinja2-time==0.2.0
jmespath==0.10.0
jsonfield==3.1.0
lazy-object-proxy==1.4.3
libsass==0.20.1
livereload==2.6.3
lxml==4.6.1
MarkupSafe==1.1.1
mccabe==0.6.1
MechanicalSoup==0.12.0
msgpack==1.0.2
msgpack-python==0.5.6
mypy==0.812
mypy-extensions==0.4.3
nodeenv==1.5.0
oauthlib==3.1.0
packaging==20.8
parse==1.18.0
parso==0.8.1
pathspec==0.8.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.1.0
pluggy==0.13.1
poyo==0.5.0
pre-commit==2.10.1
prices==1.1.0
prompt-toolkit==3.0.14
psycopg2-binary==2.8.6
ptyprocess==0.7.0
py==1.10.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.6.0
pycparser==2.20
pyee==7.0.4
pyflakes==2.2.0
Pygments==2.7.4
PyJWT==2.0.1
pylint==2.6.0
pylint-django==2.4.2
pylint-plugin-utils==0.6
pyOpenSSL==20.0.1
pyparsing==2.4.7
pyppeteer==0.2.2
pyquery==1.4.1
pytest==6.2.2
pytest-django==4.1.0
pytest-sugar==0.9.4
python-dateutil==2.8.1
python-decouple==3.4
python-slugify==4.0.1
python3-openid==3.2.0
pytz==2020.5
PyYAML==5.4.1
rcssmin==1.0.6
redis==2.10.6
regex==2020.11.13
requests==2.25.0
requests-html==0.10.0
requests-oauthlib==1.3.0
rjsmin==1.1.0
s3transfer==0.3.4
service-identity==18.1.0
six==1.15.0
snowballstemmer==2.1.0
soupsieve==2.0.1
Sphinx==3.5.1
sphinx-autobuild==2020.9.1
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==1.0.3
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.4
sqlparse==0.4.1
stripe==2.50.0
termcolor==1.1.0
testfixtures==6.17.1
text-unidecode==1.3
toml==0.10.1
tornado==6.1
tqdm==4.52.0
traitlets==4.3.3
Twisted==21.2.0
twitter-scraper==0.4.4
txaio==21.2.1
typed-ast==1.4.1
typing-extensions==3.7.4.3
urllib3==1.26.2
virtualenv==20.4.0
w3lib==1.22.0
wcwidth==0.2.5
websockets==8.1
Werkzeug==1.0.1
wrapt==1.12.1
zipp==3.4.0
zope.interface==5.2.0
